refactor(hive_event_bus): promote to directory, split tests (#1266)
- src/hive_event_bus.rs: 809 → src/hive_event_bus/mod.rs (320 LOC)
- New src/hive_event_bus/tests.rs (489 LOC, follow-up split needed)

Pure refactor: 22 tests pass.
